Move galaxy.model-dependent runtime tests to test/unit/app/tools/.
test_runtime.py imported galaxy.model and galaxy.tools.runtime which
aren't deps of the tool_util package. Split: pure tool_util_models
tests stay in test/unit/tool_util/, collection_to_runtime tests move
to test/unit/app/tools/ where galaxy.model is available.
